Add shots, use explicit cmdline args (#31)
* Add shots, use explicit cmdline args

This change adds the ability to specify a number of shots for execution when running the QIR program, which should increase speed and usability vs invoking the exe each time.

This also changes the way cmdline arguments are processed, requiring the `-f` or `--file` cmdline parameter when specifying the file to run, and `-e` or `--entrypoint` for the optional entry point name. This is a breaking change vs the positional argument pattern qir-runner used before.