Add rngseed cmdline parameter (#32)
* Add rngseed cmdline parameter

This change adds the `--rngseed` or `-r` cmdline parameter to the runner that can allow callers to seed the thread-local pseudo-random number generator used during quantum simulation. This can be handy for testing purposes by increasing predictability. Note that the seed is set outside of the shot loop, so that multiple shots should still be distinct but the sequence of results for the shots should always be the same for the same seed.

Also note that this only sets the seed used for quantum randomness. The classical randomness provided by `__quantum__qis__drawrandomint__body` and `__quantum__qis__drawrandomdouble__body` in the standard library do not use the seeded rng and remain non-deterministically random.

* Adding runner command line usage to readme

---------

Co-authored-by: Ian Davis <Ian.F.Davis@microsoft.com>