Rework LLVM-IR range support feature
This change refactors the feature flag for LLVM IR based range support to be an additive flag rather than a disabling flag, and keeps it out of the default list. That allows qir-runner (or even other consumers of qir-backend) to update the feature flag list to add it back in, and if they do not, correctly skips bulding the llvm-tools build dependency.