Update benches to compare against main
This change updates the benches section of the CI workflow to include comparison against main. Note that benches do not succeed or fail based on any specific perf boundaries, it just provoides more info in the logs. Rather than having just the latest perf values, it will show the latest as compared to the perf recorded from main.