Fix `unmap_pauli` for Pauli-Y measurement
This change fixes measurement in the Pauli-Y basis by resolving a but in how basis changes are unmapped after a call to `__quantum__qis__measure__body`. It also introduces a scenario test that verifies the behavior of Pauli-Y measurement using basis changes, assertions, and repeated measurement. Fixes Measurement in Pauli-Y basis leaves qubit in incorrect state #48.