Update rotations to use alternate phase convention
This change updates the behavior of the qir-backend `__quantum__qis__rz__body` to respect an alternate global phase convention, namely that it avoids introducing the phase of `i` in the quantum state. To make this approach consistent across other gates, the internal `mcrotation` function now also incorporates phase in the special case where the rotation matrix entries are nearly zero.

Fixes #62