Clean up sparse state once all qubits are released (#66)
This change updates the behavior of the sparse simulator in qir_backend to clean up the internal sparse representation of quantum state when all qubits are released. Specifically, if the qubit id map is empty, we know there thare no outstanding qubits being simulated, so the state can be reset back to a perfect ground state of `|0⟩ = 1 + 0i`. This lets the simulation drop any accumulated phase that should no longer be preserved since no qubits are "active" to carry it.