Fix panic in dump reordering after swap (#67)
This fixes an issue where state dump could crash if run after a swap of qubit ids. Specifically, a qubit id swap followed by a release of one of the qubits can put the id map into a state where not every desired id for the swap is occupied by an existing qubit. As a result, updating the map for the swapped key should be a best effort and not a panic if the key is not present.  Also adds a unit test to cover this specific case.