Preserve queued rotations across id swap
This change fixes an issue with `swap_qubit_ids` that could cause queued rotations on swapped qubits to get lost. The updated logic correctly queries the maps before applying changes and updates the right qubits with the right values. It also introduces a test case that verifies swapping respects queued rotations, and slightly optimizes the queuing of `rx` rotations.