More static result support in measurement API
This change updates the signature of `__quantum__qis__mresetz__body` and adds it to the set of linked functions so that it can be used by programs expecting it as an intrinsic that uses static result allocation. This also adds support for two versions of `__quantum__rt__m__body` during linking to allow for both dynamic and static result allocation.