Fix multicontrolled Ri, tests
The test utility for comparing simulator functions against matrix applications was not using enough qubits to fully verify the resulting state, and fixing this revealed a bug in the way `mcphase` was used to implement a multicontrolled Ri rotation (the phase was not applied globally). This change updates the tests and fixes `mcphase` to apply to all states instead of just states with a 1 in the computational basis.