Support additional operation queuing
This adds more operation queuing that supports X, Y, Z, S, Sadj, T, and Tadj (along with controls) when feasible. When flushed, the queue will apply each of these operations in one iteration through the state vector rather than iterating the state vector for each. In practice, this seems to provide small benefits for programs will few qubits and thus small vectors (6% to 7%) but the benefit scales with the size of the state vector (observed 23% to 50% on larger programs).