Use `IndexMap` for performance boost in qubit allocation (#215)
* Use `IndexMap` for performance boost in qubit allocation

This change pulls in the `IndexMap` struct from https://github.com/microsoft/qsharp/tree/main/compiler/qsc_data_structures to use for the qubit id mapping, rather than `FxHashMap`. In most use cases qubit ids tend to be dense and sequential, making `IndexMap` an optimal fit. This also cuts down on the execution time of the `test_large_state` unit test by an order of magnitude, and the new benchmark shows significant gains as well. Other benchmaks showed a small decrease in the perf (around 5%) but since those are measured in microseconds it seems to be a worthwile trade-off (for example, the X gate benchmark was slower by about 0.15 microsecods).

* Reduce bench size to 2k qubits