Add Rayon parallelism, prefer vector over hash map
This updates the sparse state simulator to use parallism via the rayon crate, further improving performance in environments that support multi-threading.
Part of enabling parallelism is changing the `SparseState` data structure to be a simple vector of unordered tuples, since most of the time we just iterate through all the states. The only gates that need the expensive hash table are the (already expensive) H, Rx, and Ry. Now those gates convert into hash tables internally to facilitate the lookups needed.