Use simpler data structure for better perf on gates
This updates the `SparseState` data structure to be a simple vector of unordered tuples, since most of the time we just iterate through all the states. The only gates that need the expensive hash table are the (already expensive) H, Rx, and Ry. Now those gates convert into hash tables internally to facilitate the lookups needed. In testing, this leads to perf improvements around 20%.