Prefer vector over hash map (#239)
* Prefer vector over hash map

This changes the `SparseState` data structure to be a simple vector of unordered tuples, since most of the time we just iterate through all the states. The only gates that need the expensive hash table are the (already expensive) H, Rx, and Ry. Now those gates convert into hash tables internally to facilitate the lookups needed.

* Remove commented line, extra newline

* Optimize measurement and release to avoid extra iteration through state

* PR feedback